Allow phones to see 8 lockscreen targets
The frameworks_base was modified to allow phones to have a total of 8 targets, like tablets.
While the lockscreen itself seemed fine, the layout in settings was shifted to the bottom.
This centers the GlowPadView so all 8 targets can be seen and set.
Not the best fix in the world, but I'm still working on it.