Vectorize Screw'd Settings icon for themes
Since everyone seems to ask me why this icon isn't "Theme-able", and people
don't seem to understand why this icon isn't changed when a theme is applied,
I've made this into a vector, and set the color to take it's cue from whatever
the current system color is. By default it's purple since we're Screw'd, but
if a theme changes the system color, this icon will change to match.

So the icon is "theme-able", happy now?!?!?!?